requirements.txt
# ============================================================
#  Virex Security System - Requirements
#  Install: pip install -r requirements.txt
# ============================================================

# --- Web Framework ---
flask==2.2.5
flask-cors==4.0.0
flask-limiter==3.5.0

# --- Authentication ---
PyJWT==2.8.0

# --- AI & Chatbot ---
google-generativeai==0.8.3

# --- Machine Learning ---
scikit-learn==1.3.0
numpy==1.24.3
pandas==1.5.3
joblib==1.3.2

# --- ML Training (for train_model_enhanced.py) ---
matplotlib==3.7.2
seaborn==0.12.2

# --- HTTP Requests ---
requests==2.31.0

# --- Utilities ---
python-dotenv==1.0.0
psutil==5.9.5

# --- Testing ---
pytest==7.4.0
pytest-flask==1.2.0
coverage==7.3.0

# --- Production Server ---
gunicorn==21.2.0